requirements.txt
fastapi==0.121.3
uvicorn==0.38.0  
python-dotenv==1.2.1 
requests==2.32.5